Added matans fileuploader to git

+++ fileuploader/src/main/java/org/mila/uploader/UploadAction.java
@@ -3,8 +3,13 @@ package org.mila.uploader;
 import java.io.File;
 import java.io.IOException;
 import java.sql.Timestamp;
+import java.util.Collections;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipException;
+import java.util.zip.ZipFile;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.filefilter.MagicNumberFileFilter;
 import org.apache.log4j.Logger;
 import org.mila.uploader.entities.TagRequest;
 import org.mila.uploader.entities.TagRequestState;
@@ -28,6 +33,9 @@ public class UploadAction extends ActionSupport {
 	private String tempDir;
 	private UserService userService;
 	private TagRequestService tagRequestService;
+	private static byte[] ZIP_MAGIC = new byte[] { 0x50, 0x4B, 0x03, 0x04 };
+	private static MagicNumberFileFilter zipfilter = new MagicNumberFileFilter(
+			ZIP_MAGIC);
 
 	@Autowired
 	public void setUserService(UserService userService) {
@@ -68,27 +76,71 @@ public class UploadAction extends ActionSupport {
 	public String execute() {
 		User user = userService.find((String) ActionContext.getContext()
 				.getSession().get("logged-in-user"));
+
+		if (file == null || filename == null)
+			return ERROR;
+		
+		if (zipfilter.accept(file)) {
+			/* This is a zip file */
+			logger.info("Processing zipfile upload, tempDir is: " + tempDir);
+			ZipFile zipfile;
+			try {
+				zipfile = new ZipFile(file);
+			} catch (ZipException e) {
+				logger.error("Some sort of zip file problem, giving up", e);
+				return ERROR;
+			} catch (IOException e) {
+				logger.error("Some sort of zip file IOException, giving up", e);
+				return ERROR;
+			}
+
+			/* If we got this far in the code, we don't actually fail anymore... */
+			for (ZipEntry entry : Collections.list(zipfile.entries())) {
+				String uploadedFilename = filename + ":" + entry.getName();
+				try {
+					assert tempDir != null;
+					File copy = File.createTempFile("input", "", new File(
+							tempDir));
+					FileUtils.copyInputStreamToFile(
+							zipfile.getInputStream(entry), copy);
+					copy.deleteOnExit();
+					createTagRequest(user, copy, uploadedFilename);
+				} catch (IOException e) {
+					logger.warn(
+							"Couldn't create tag request for a zip entry, continuing...",
+							e);
+				}
+			}
+			return SUCCESS;
+		} else {
+			/* This is just a normal file */
+			logger.info("Processing upload, tempDir is: " + tempDir);
+			try {
+				assert tempDir != null;
+				File copy = File.createTempFile("input", "", new File(tempDir));
+				FileUtils.copyFile(file, copy);
+				copy.deleteOnExit();
+				return createTagRequest(user, copy, filename);
+			} catch (IOException e) {
+				logger.error("Can't save uploaded file, giving up tagging", e);
+				return ERROR;
+			}
+		}
+	}
+
+	private String createTagRequest(User user, File workfile,
+			String uploadedFilename) {
 		TagRequest tagreq = new TagRequest();
 		tagreq.setUser(user);
 		tagreq.setState(TagRequestState.WAITING);
-		tagreq.setUploadedFilename(filename);
+		tagreq.setUploadedFilename(uploadedFilename);
 		tagreq.setTimestamp(new Timestamp(System.currentTimeMillis()));
 		tagRequestService.save(tagreq);
 
 		logger.debug("Adding a tag request, the content-type was: "
 				+ contentType);
 
-		try {
-			assert tempDir != null;
-			logger.info("Processing upload, tempDir is: " + tempDir);
-			File copy = File.createTempFile("input", "", new File(tempDir));
-			FileUtils.copyFile(file, copy);
-			copy.deleteOnExit();
-			asyncBean.handle(copy, tagreq);
-		} catch (IOException e) {
-			logger.error("Can't save uploaded file, giving up tagging", e);
-			return ERROR;
-		}
+		asyncBean.handle(workfile, tagreq);
 
 		return SUCCESS;
 	}
